feat: extract the library and fix spell
