+++ spx-gui/src/components/top-menu/TopMenu.vue
@@ -1,7 +1,7 @@
 <!--
  * @Author: Xu Ning
  * @Date: 2024-01-12 16:52:20
- * @LastEditors: xuning 453594138@qq.com
+ * @LastEditors: Xu Ning
  * @LastEditTime: 2024-01-17 21:26:30
  * @FilePath: /spx-gui-front-private/src/components/top-menu/TopMenu.vue
  * @Description:
