+++ spx-gui/src/language/index.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: Yao xinyue kother@qq.com
  * @Date: 2024-01-18 01:56:51
- * @LastEditors: xuning 453594138@qq.com
+ * @LastEditors: Xu Ning
  * @LastEditTime: 2024-01-17 21:26:53
  * @FilePath: src/language/index.ts
  * @Description:
